Speed up AppWindow.ProcessTitle for UWP also, & refactor new code
